Fix ambiguous operators under C++20 (#5567)
When resolving `operator!=`, C++20 also considers `operator==` implementations
when the types on `operator!=` do not match exactly. This caused the modified
code to have no most-specific overload to choose, resulting in an error. This is
actually a bug in the language that is being fixed, but there exist compilers
without the fix applied.

Work around the problem by updating the types in the declaration of `operator==`
and `operator!=` to be more exact.

This is a copy of #5029 with formatting fixes.