Ensure a deterministic order in the type names section (#5590)
Before this PR we iterated over an unordered set. Replace that with an
iteration on a vector. (Also, the value in the set was not even used, so
this should even be faster.)

Add random names in the fuzzer to types, the lack of which is I believe
the reason this was not detected before.