Fix a crash in MemoryPacking due to an unreachable pointer (#5623)
Previously, the pointer type for newly emitted instructions was determined by
the type of the destination pointer on a memory.init instruction, but that did
not take into account that the destination pointer may be unreachable. Properly
look up the pointer type on the memory instead to fix the problem.

Fixes #5620.