[Wasm GC] Improve GC operation coverage by using locals more (#5661)
When we emit e.g. a struct.get's reference, this PR makes us prefer a non-nullable
value, and even to reuse an existing local if possible. By doing that we reduce
the risk of a trap, and also by using locals we end up testing operations on the
same data, like this:

x = new A();
x.a = ..
foo(x.a)

In contrast, without this PR each of those x. uses might be new A().