[Wasm GC] OptimizeInstructions: Don't turn ref.test into unreachable immediately (#5673)
Emit an unreachable, but guarded by a block as we do in other cases in this pass, to avoid
having unreachable code that is not fully propagated during the pass (as we only do a full
refinalize at the end). See existing comments starting with
"Make sure to emit a block with the same type as us" in the pass.

This is mostly not a problem with other casts, but ref.test returns an i32 which we have
lots of code that tries to optimize.