Fallback to direct inlining if the outline will be inlined. (#5698)
This workarounds the extra work around the edge case where;

 - Function is too big to full-inline
 - It is a candidate for partial inline
 - Outlined version becomes eligible for full-inline.

In such a case, binaryen would introduce a temporary state with
partial inlined functions and later on inline them. J2CL hit this
scenario for String literal which resulted in significant
regressions in compilation time.

This patch updates partial inlining analysis to identify the
edge case and direct to full-inlining when that happens.