Move casts which are immediate children of local.gets to earlier local.gets (#5744)
In the OptimizeCasts pass, it is useful to move more refined casts as early as
possible without causing side-effects. This will allow such casts to potentially
trap earlier, and will allow the OptimizeCasts pass to use more refined casts
earlier. This change allows a more refined cast to be duplicated at an earlier
local.get expression. The later instance of the cast will then be eliminated in
a later optimization pass.

For example, if we have the following instructions:

(drop
  (local.get $x)
)
(drop
  (ref.cast $A
    (local.get $x)
)
(drop
  (ref.cast $B
    (local.get $x)
  )
)

Where $B is a sublcass of $A, we can convert this to:

(drop
  (ref.cast $B
    (local.get $x)
  )
)
(drop
  (ref.cast $A
    (local.get $x)
)
(drop
  (ref.cast $B
    (local.get $x)
  )
)

Concretely we will save the first cast to a local and use it in the other
local.gets.