[EH] Add pass to remove EH instructions (#5770)
This pass strips all EH stuff, including EH instructions and tags, from
the input module and disables the EH feature from the features section.

1. This removes `catch` and `catch_all` blocks from the code. So
```wast
(try
  (do
    (some code)
  )
  (catch
    ...
  )
)
```
becomes just `(some code)`. Note that all `rethrow`s will be removed
with `catch`es. Note that all `rethrow`s will be removed with `catch`es.

2. This converts 'throw (...)` into `unreachable`. Note that `rethrows

3. This removes all tags from the module, which are unused anyway after
1 and 2.

4. This removes exception handling feature from the features section.

You can use the pass with
```console
$ wasm-opt --enable-exception-handling --strip-eh INPUT -o OUTPUT
```
This is not an optimization pass, so it is not run unless you specify
the pass explicitly.

This is in effect similar to Clang's `-fignore-exceptions`, in which you
can throw but it will result in a crash and we compile away all landing
pads. This can be used for people who don't (or can't) use
`-fignore-exceptions` in their build settings or who want to compile
away `catch` blocks later.

Closes emscripten-core/emscripten#19585.