[NFC] Simplify `Tuple` by making it an alias of `TypeList` (#5775)
Rather than wrap a `TypeList`, make `Tuple` an alias of `TypeList`. This means
removing `Tuple::toString`, but that had no callers and was of limited use for
debugging anyway. In return, the use of tuples becomes much less verbose.

In the future, it may make sense to remove one of `Tuple` and `TypeList`.