Make the Transfer Function a Generic, Interchangeable Type for the Monotone Analyzer (#5794)
This change makes the transfer function an object separate from the monotone analyzer. The transfer function class type is a generic template of the monotone analyzer, and an instance of the transfer function is instantiated and then passed into the monotone analyzer via its constructor. The API of the transfer function is enforced by a static assertion.

This change also introduces LivenessTransferFunction as a transfer function for liveness analysis as an example.