TypeRefining: Add casts when we must (#5840)
See the example in the code and test for a situation that requires this for validation.

To fix validation we add a cast. That should practically always be removed by later
optimizations, and the fact it took the fuzzer this long to even find such a situation
also adds confidence that this won't be adding overhead (and in this situation, the
optimizer will definitely remove the cast).