OptimizeCasts: Connect adjacent blocks in LinearExecutionWalker (#5866)
Followup to #5860, this does the same for (part of) OptimizeCasts.

As there, this is valid because it's ok if we branch away. This part of the pass
picks a different local to get when it knows locals have the same values but one
is more refined. It is ok to add a tee earlier even if it isn't used later.