[typed-cont] Allow result types on tags (#5997)
This PR is part of a series that adds basic support for the typed continuations proposal.

This PR relaxes the restriction that tags must not have results , only params. Tags with
results must not be used for exception handling and are only allowed if the typed
continuations feature is enabled.

As a minor point, this PR also changes the printing of tags without params: To make the
presentation consistent, (param) is omitted when printing a tag.