Update IRBuilder to visit control flow correctly (#6124)
Besides If, no control flow structure consumes values from the stack. Fix a
bug in IRBuilder that was causing it to pop control flow children. Also fix a
follow on bug in outlining where it did not make the If condition available on
the stack when starting to visit an If. This required making push() part of
the public API of IRBuilder.

As a drive-by, also add helpful debug logging to IRBuilder.

Co-authored-by: Ashley Nelson <nashley@google.com>