wasm-metadce: Improve name deduplication (#6138)
Avoid adding suffixes when we don't need them to keep names unique.

As background, the suffixes are not used by emcc at all, so they are just
for internal use in the tool. How that works is that metadce gets as input
the list of things the user cares about, with names for them, so it knows
the proper names to give imports and exports, and makes up names for
other things. Those made up names will not be read by the user, so we
can make them prettier as this PR does without breaking anything.

The main benefit of this PR is to make debugging easier.