wasm-reduce: Improve tryToReduceCurrentToConst() (#6193)
Avoid replacing with the exact same thing in the case of RefNull and a default
tuple.

Also be more careful with handling of numbers. Before we exited immediately
if we saw a number, but we can try to replace a number with a 0 or a 1, even
if it was a number before. That is, we consider 1 simpler than e.g. 12345678, and
0 simpler than 1.