StringLowering: Use an array16 type in its own rec group (#6302)
The input module might use an array of 16-bit elements type that is somewhere in a
giant rec group, but that is not valid for imported strings: that array type is now on an
import and must match the expected ABI, which is to be in its own personal rec group.

The old array16 type remains in the module after this transformation, but all uses of it
are replaced with uses of the new array16 type.

Also move makeImports to after updateTypes: there are no types to update in the new
imports. That does not matter but it can make debugging less pleasant, so improve it.