Fuzzer: Remove Asyncify integration (#6309)
Fuzzing Asyncify has a significant cost both in terms of the complexity in
the fuzzer and the slowness of the fuzzing. In practice it was useful years ago
when Asyncify was written but hasn't found anything for a while, and Asyncify
is really deprecated given JSPI. For all those reasons, remove it from the fuzzer.

We do still have lots of normal coverage of asyncify in lit tests, unit tests, and
the Emscripten test suite.

Removing this will also make future improvements to the fuzzer simpler.