[Emscripten port] Improve emcc flags (#6349)
No changes here to binaryen.js/wasm builds.

1. Add a flag to enable pthreads.

2. Use SINGLE_FILE on binaryen.js/.wasm as before, which is nice for library users
as they want just a single file to distribute for Binaryen support. For other builds
like wasm-opt.js etc. no longer set SINGLE_FILE, as that type of build wants to be
a replacement for a normal wasm-opt build as much as possible, so avoid the
overhead of SINGLE_FILE.

(Previously we disabled SINGLE_FILE also in the case of BUILD_FOR_BROWSER
but I don't think we need to special-case that any more.)